requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
openai
langchain-openai
anthropic
langchain-anthropic
google-generativeai
langchain-google-genai
transformers
huggingface-hub
sentence-transformers
accelerate
torch
langchain-groq
faiss-cpu
chromadb
pypdf
unstructured
python-docx
beautifulsoup4
numpy
pandas
scipy
tqdm
python-dotenv
requests
sqlalchemy
ipykernel
streamlit
grandalf
langchain-huggingface
langchain-experimental
wikipedia